requirements.txt
adbc-driver-manager==0.10.0
adbc-driver-postgresql==0.10.0
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
flask==3.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
mistune==3.0.2
numpy==1.26.4
packaging==23.2
polars-lts-cpu==0.20.10
psycopg2-binary==2.9.9
pyarrow==15.0.0
pydantic==2.6.2
pydantic-core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing-extensions==4.9.0
tzdata==2024.1
werkzeug==3.0.1
